+++ changelog.md
@@ -1,2 +1,3 @@
 ## Added
 - Bitmap font parsing ([text format of .fnt](https://www.angelcode.com/products/bmfont/))
+- Bitmap fonts cache (nsFontsCache)